{"types":"C","runs":10}

+++ problem15.java
@@ -2,10 +2,10 @@ import acm.program.*;
 
 public class problem15 extends ConsoleProgram {
 	public void run() {
-		double n1 = readInt("first Number :" );
-		double n2 = readInt("SecondNumber :" );
-		double total = n1 + n2;
-		double arithmeticMiddle = total/2;
+		int n1 = readInt("first Number :" );
+		int n2 = readInt("SecondNumber :" );
+		int total = n1 + n2;
+		int arithmeticMiddle = total/2;
 		println(arithmeticMiddle);
 	}
 }
